Split multi-class files in Kafka.Ksql.Linq
