+++ src/HoppingWindow.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace Kafka.Ksql.Linq;
+
+public static class HoppingWindow
+{
+    public static WindowDef Of(TimeSpan size) => new WindowDef().HoppingWindow().Size(size);
+    public static WindowDef OfMinutes(int minutes) => Of(TimeSpan.FromMinutes(minutes));
+}
