+++ src/Serialization/Abstractions/SerializationFormat.cs
@@ -1,0 +1,8 @@
+namespace Kafka.Ksql.Linq.Serialization.Abstractions;
+
+public enum SerializationFormat
+{
+    Avro,
+    Json,
+    Protobuf
+}
